ray again: anti aliasing

+++ C/Info1/ray_2/srcs/parser/get_cam.c
@@ -6,7 +6,7 @@
 /*   By: msarr <msarr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/30 18:59:57 by msarr             #+#    #+#             */
-/*   Updated: 2014/12/25 18:29:53 by msarr            ###   ########.fr       */
+/*   Updated: 2014/12/26 13:04:26 by msarr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ static int		cam_exit()
 void		cam_setup(t_cam *c)
 {
 	c->fov = c->focal;
-   	c->focal = -(SCREEN_W / ( 2 * tan (c->fov / 2)));
-	c->up = new(0.0f, 1.0f, 0.0f);
+   	c->focal = -((double)SCREEN_W / ( 2.0 * tan(c->fov / 2.0)));
+	c->up = new(0.0, 1.0, 0.0);
     c->dir = sub(c->lookat, c->pos);
    	c->dir = normal(c->dir);
    	c->screen = new(SCREEN_W, SCREEN_H, c->focal);
@@ -37,8 +37,8 @@ void		cam_setup(t_cam *c)
 	c->up = cross(c->dir, c->upright);
 	// Maintenant nous avons toutes les informations pour détérminer la position 
 	// en haut à gauche du viewplane.
-	c->upleft = add(c->pos, sub(add(mult2(c->dir, c->focal), mult2(c->up, (SCREEN_H/2.0f))),
-		mult2(c->upright, (SCREEN_W/2.0f))));
+	c->upleft = add(c->pos, sub(add(mult2(c->dir, c->focal), mult2(c->up, ((double)SCREEN_H / 2.0))),
+		mult2(c->upright, ((double)SCREEN_W / 2.0))));
 }
 
 t_lex			*get_cam(t_cam *cam, t_lex *lex)
